finish the server. check the delete post function, she stack on the result

+++ server/db/todos.js
@@ -1,32 +1,44 @@
 const pool = require("./pool");
 
-async function getTodos(id) {
+async function getTodos(userId) {
     const todosQuery = `
         SELECT * FROM todos
         WHERE userId = ?
     `;
-    const [data] = await pool.query(todosQuery, [id]);
-    return data;
+    const [data] = await pool.query(todosQuery, [userId]);
+    if (data.length > 0) {
+        return data;
+    } else {
+        throw new Error("please provide the correct user id");
+    }
 };
 
-async function getCompletedTodos(id) {
+async function getCompletedTodos(userId) {
     const todosQuery = `
         SELECT * FROM todos
         WHERE userId = ?
         AND completed = 1
     `;
-    const [data] = await pool.query(todosQuery, [id])
-    return data;
+    const [data] = await pool.query(todosQuery, [userId])
+    if (data.length > 0) {
+        return data;
+    } else {
+        throw new Error("please provide the correct user id");
+    }
 };
 
-async function getUnCompletedTodos(id) {
+async function getUnCompletedTodos(userId) {
     const todosQuery = `
         SELECT * FROM todos
         WHERE userId = ?
         AND completed = 0
     `;
-    const [data] = await pool.query(todosQuery, [id])
-    return data;
+    const [data] = await pool.query(todosQuery, [userId])
+    if (data.length > 0) {
+        return data;
+    } else {
+        throw new Error("please provide the correct user id");
+    }
 };
 
 async function addTodo(userId, title) {
@@ -35,20 +47,46 @@ async function addTodo(userId, title) {
         VALUES (?, ?, 0)
     `;
     const [result] = await pool.query(insertQuery, [userId, title]);
-    return result;
+    if (result.affectedRows > 0) {
+        return result;
+    } else {
+        throw new Error("add failed! check the details and try again");
+    }
 };
 
-async function updateTitle(id, title) {
+async function updateTitle(userId, todoId, title) {
+    const ownerCheckQuery = `
+        SELECT userId 
+        FROM todos 
+        WHERE id = ?;
+    `;
+    const [ownerCheckResult] = await pool.query(ownerCheckQuery, [todoId]);
+
+    if (ownerCheckResult.length === 0 || ownerCheckResult[0].userId !== userId) {
+        throw new Error("You do not have permission to update this todo");
+    }
     const sql = `
         UPDATE todos
         SET title = ?
         WHERE id = ?
     `
-    const update = await pool.query(sql, [title, id])
-    return update;
+    const [update] = await pool.query(sql, [title, id])
+    if (update.affectedRows > 0) {
+        return update;
+    } else {
+        throw new Error("Todo not found or title unchanged");
+    }
 };
 
-async function updateCompleted(id) {
+async function updateCompleted(userId, todoId) {
+    const ownerCheckQuery = `
+        SELECT userId FROM todos WHERE id = ?;
+    `;
+    const [ownerCheckResult] = await pool.query(ownerCheckQuery, [todoId]);
+    if (ownerCheckResult[0].userId !== userId) {
+        throw new Error("You do not have permission to update this todo");
+    }
+    
     const sql = `
         UPDATE todos
         SET completed = CASE
@@ -57,17 +95,34 @@ async function updateCompleted(id) {
         END
         WHERE id = ?
     `
-    const result = await pool.query(sql, [id])
-    return result;
+    const [result] = await pool.query(sql, [todoId])
+    if (result.affectedRows > 0) {
+        return result;
+    } else {
+        throw new Error("Todo not found or completed unchanged")
+    }
 };
 
-async function deleteTodo(id) {
+async function deleteTodo(userId, todoId) {
+    const ownerCheckQuery = `
+        SELECT userId 
+        FROM todos 
+        WHERE id = ?;
+    `;
+    const [ownerCheckResult] = await pool.query(ownerCheckQuery, [todoId]);
+    if (ownerCheckResult[0].userId !== userId) {
+        throw new Error("You do not have permission to delete this todo");
+    }
     const deleteQuery = `
         DELETE FROM todos
         WHERE id = ?
     `
-    const result = await pool.query(deleteQuery, [id])
-    return result;
+    const [result] = await pool.query(deleteQuery, [todoId])
+    if (result.affectedRows > 0) {
+        return result;
+    } else {
+        throw new Error("Todo not found or not deleted");
+    }
 };
 
 module.exports = {
